[Structured metrics] Use int64_t instead of int
The structured metrics spec is to use int64 for int metrics, but the
code currently uses `int`. This CL replaces it with `int64_t`.

This should be a backwards compatible change, and the proto used for
upload already uses int64_t.

Bug: 1148168
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2885410
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#881428}

+++ components/metrics/structured/event_base.cc
@@ -28,7 +28,7 @@ void EventBase::AddStringMetric(uint64_t name_hash, const std::string& value) {
   metrics_.push_back(metric);
 }
 
-void EventBase::AddIntMetric(uint64_t name_hash, int value) {
+void EventBase::AddIntMetric(uint64_t name_hash, int64_t value) {
   Metric metric(name_hash, MetricType::kInt);
   metric.int_value = value;
   metrics_.push_back(metric);
